aaaaaaaa

+++ code/modules/antagonists/paradox_clone/paradox_clone.dm
@@ -83,7 +83,13 @@
 	if(!target?.current)
 		explanation_text = "Free Objective"
 		CRASH("WARNING! [ADMIN_LOOKUPFLW(owner)] paradox clone objectives forged without an original!")
-	explanation_text = "Murder and replace [target.name], the [!target_role_type ? target.assigned_role.title : english_list(target.get_special_roles())]. Remember, your mission is to blend in, do not kill anyone else unless you have to!"
+	// BUBBER EDIT ADDITION BEGIN - No RR for non-RR targets
+	var/datum/preferences/prefs = target.current.client?.prefs
+	if (isnull(prefs) || !prefs.read_preference(/datum/preference/toggle/be_round_removed)) // the nullcheck will realistically never trigger
+		explanation_text = "Replace [target.name], the [!target_role_type ? target.assigned_role.title : english_list(target.get_special_roles())]. Remember, you're not allowed to remove them from the round, so you'll have to convince everyone you're the real one - strike a deal - or something else!"
+	else
+		explanation_text = "Murder and replace [target.name], the [!target_role_type ? target.assigned_role.title : english_list(target.get_special_roles())]. Remember, your mission is to blend in, do not kill anyone else unless you have to!"
+	// BUBBER EDIT ADDITION END
 
 ///Static bluespace stream used in its ghost poll icon.
 /obj/effect/bluespace_stream
